+++ tools/xdhq/proxy/LIB/xdhqxdh.cpp
@@ -920,7 +920,6 @@ namespace {
 		}
 		void Log_( const str::dString &Message )
 		{
-			return;
 		qRH;
 			logq::rLogRack<> Log;
 		qRB;